team-management-api/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2024.1
urllib3==2.4.0
